fix AuthPacket Wrong

+++ common/src/main/java/io/mycat/beans/mysql/packet/MySQLPayloadReadView.java
@@ -32,7 +32,7 @@ public interface MySQLPayloadReadView {
 
     long readFixInt(int length);
 
-    int readLenencInt();
+    Long readLenencInt();
 
     String readFixString(int length);
